suppurt delete list of topics

+++ fortscale/fortscale-utils/src/main/java/fortscale/utils/kafka/KafkaUtils.java
@@ -8,6 +8,8 @@ import kafka.utils.ZkUtils;
 import org.I0Itec.zkclient.ZkClient;
 import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Value;
+import scala.collection.Seq;
+
 import java.util.Arrays;
 
 import java.io.File;
@@ -44,17 +46,37 @@ public class KafkaUtils extends CleanupDeletionUtil {
         Object[] objectsArray = topics.toArray();
         String[] topicsArray = Arrays.copyOf(objectsArray, objectsArray.length, String[].class);
 
-        TopicCommand.TopicCommandOptions opts = new TopicCommand.TopicCommandOptions(topicsArray);
-        try {
-            TopicCommand.deleteTopic(zkClient, opts);
-        } catch (AdminOperationException ex) {
-            logger.error("failed to drop all {} topics, {}", topics.size(), ex.getMessage());
-            logger.error(ex.toString());
-            return false;
-        }
 
-        logger.info("dropped all {} topics", topics.size());
-        return true;
+        if (isBrutalDelete) {
+            numberOfTopicsDeleted = 0;
+            for (String topic: topics) {
+                if (deleteTopic(topic, zkClient, doValidate)) {
+                    numberOfTopicsDeleted++;
+                }
+            }
+            zkClient.close();
+            if (numberOfTopicsDeleted == topics.size()) {
+                logger.info("dropped all {} topics", topics.size());
+                return true;
+            }
+            logger.error("failed to drop all {} topics, dropped only {}", topics.size(), numberOfTopicsDeleted);
+            return false;
+        } else {
+
+            //delete using API
+            TopicCommand.TopicCommandOptions opts = new TopicCommand.TopicCommandOptions(topicsArray);
+            try {
+                TopicCommand.deleteTopic(zkClient, opts);
+                zkClient.close();
+            } catch (AdminOperationException ex) {
+                logger.error("failed to drop all {} topics, {}", topics.size(), ex.getMessage());
+                logger.error(ex.toString());
+                return false;
+            }
+
+            logger.info("dropped all {} topics", topics.size());
+            return true;
+        }
 
     }
 
@@ -95,7 +117,8 @@ public class KafkaUtils extends CleanupDeletionUtil {
         logger.debug("establishing connection to zookeeper");
         ZkClient zkClient = new ZkClient(zookeeperConnection, zookeeperTimeout);
         logger.debug("connection established, fetching topics");
-        Collection<String> topics = scala.collection.JavaConversions.seqAsJavaList(ZkUtils.getAllTopics(zkClient));
+        Collection<String> topics = scala.collection.JavaConversions.seqAsJavaList((Seq<String>)ZkUtils.
+                getAllTopics(zkClient));
         zkClient.close();
         return topics;
     }
@@ -110,10 +133,17 @@ public class KafkaUtils extends CleanupDeletionUtil {
     @Override
     public boolean deleteAllEntities(boolean doValidate) {
         Collection<String> topics = getAllEntities();
+        ZkClient zkClient = new ZkClient(zookeeperConnection, zookeeperTimeout);
         boolean success = false;
         logger.debug("found {} topics to delete", topics.size());
         if (isBrutalDelete) {
             //delete physical files
+            for (String topic: topics) {
+                //delete zookeeper entries first
+                String topicPath = ZkUtils.getTopicPath(topic);
+                logger.debug("attempting to delete topic {}", topic);
+                zkClient.deleteRecursive(topicPath);
+            }
             success = cleanKafakDataFolders(doValidate);
         } else {
             //delete using API
@@ -135,17 +165,12 @@ public class KafkaUtils extends CleanupDeletionUtil {
             logger.error("no kafka data folder {} found", kafkaDataFolder);
             return false;
         }
-        /*try {
-            //FileUtils.cleanDirectory(directory);
-        } catch (IOException ex) {
-            logger.error("failed to clean folder {} - {}", directory.getAbsolutePath(), ex);
-            return false;
-        }*/
+
 
         String[] cmdArray = {"bash", "-c", "sudo rm -rf /var/local/kafka/data"};
         try {
             Process runCmd = Runtime.getRuntime().exec(cmdArray);
-            System.out.println("Command Executed Successfully");
+            logger.info("Command Executed Successfully");
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -176,11 +201,13 @@ public class KafkaUtils extends CleanupDeletionUtil {
         for(String folderName : folders) {
             File folder = new File(kafkaDataFolder + "/" + folderName);
             if (folderName.startsWith(prefix) && folder.isDirectory()) {
+
+                String[] cmdArray = {"bash", "-c", "sudo rm -rf " + kafkaDataFolder + "/" + folderName};
                 try {
-                    FileUtils.deleteDirectory(folder);
-                } catch (IOException ex) {
-                    logger.error("failed to delete kafak data folder {} - {}", folder.getAbsolutePath(), ex);
-                    return false;
+                    Process runCmd = Runtime.getRuntime().exec(cmdArray);
+                    logger.info("Command Executed Successfully");
+                } catch (IOException e) {
+                    e.printStackTrace();
                 }
                 if (validate && folder.exists()) {
                     logger.error("failed to delete kafak data folder {}", folder.getAbsolutePath());
